feat:集成Mybatis官方代码生成器

+++ member/src/main/java/com/sdu/train/member/controller/MemberController.java
@@ -14,7 +14,7 @@ public class MemberController {
     private MemberService memberService;
 
     @GetMapping("/count")
-    public Integer count() {
+    public Long count() {
         return memberService.count();
     }
 }
